move functions out of db.inc.php, move config to file not present in repo, add template
