+++ scripts/config.inc.php.template
@@ -1,0 +1,8 @@
+<?php
+$tile_size = 0.01;
+$frontend_url = $_SERVER['REQUEST_SCHEME'] . '://' . $_SERVER['HTTP_HOST'] . '/whodidit/';
+$db_hostname = 'localhost';
+$db_database = 'wdi';
+$db_username = 'wdi';
+$db_password = '';
+?>
